+++ backend/src/main/java/com/cogumelos/dto/formulacao/FormulacaoResponse.java
@@ -18,7 +18,7 @@ import java.util.List;
 public record FormulacaoResponse(String id, String nome, String usuarioId, String usuarioNome,
                                  String especieId, String especieNome, Double cnTotal, Double phMedio, String criadoEm,
                                  String status, boolean cnDentroFaixa, List<FormulacaoInsumoResponse> insumos,
-                                 Double umidadeDesejada, Double pesoBlocoKg, Integer totalBlocos) {
+                                 Double umidadeDesejada, Double pesoBlocoKg, Double totalBlocos) {
     public static FormulacaoResponse from(Formulacao f) {
         return new FormulacaoResponse(f.getId(), f.getNome(), f.getUsuario().getId(), f.getUsuario().getNome(),
                 f.getEspecie().getId(), f.getEspecie().getNome(), f.getCnTotal(), f.getPhMedio(),
